stowage-optimizer/requirements.txt
ortools==9.11.4210
pymongo==4.10.1
pandas==2.2.3
openpyxl==3.1.2
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
